Add the modfetch package
Add the modfetch package, implementing a Load function that can be used
to fetch and cache modules.
The moduleJSON type has been copied from the go mod help download
documentation, but the exported Module type uses the ModuleError type
for the Error field.

Add the modfetch command to test the modfetch package.