invoke: don't include stderr if empty
go mod download -json does not write anything on stderr in case of
errors (excluding debug messages with go1.13).  With the current
implementation, the error returned by modfetch is something like
  "modfetch: load: go mod download -json <mod@ver>: exit status 1: "

In the Go function, ensure that the stderr content is not included in
the error, if it is empty.