modfetch: collect all the errors in Error.Stderr
The go mod download -json command does not print errors on stderr, like
go list -json; instead it stores them in the Module.Error field.
This make it impossible to the caller of Load to report the errors to
the user.

When invoke.Go returns an error, decode and scan the packages from
stdout, collect the error messages and store the data in Error.Stderr.

Note that what we do is the exact opposite of
golang.org/x/tools/go/packages, where it scans stderr and stores the
error messages in the Package.Error field.