gocmd: fix check on command line arguments
In ee1c959 (cmd/gocmd: make debugging optional), the flag package was
used to parse command line arguments, but the code still checked
len(os.Args) instead of flag.NArg().  This caused the gocmd to panic if
the -debug flag was specified but there were no arguments.

Update the main fuction to exit if flag.NArg() is 0.